consultas de usuarios junto con rh

+++ database/migrations/2021_02_10_160558_create_users_table.php
@@ -15,7 +15,7 @@ class CreateUsersTable extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->integer('id_rhh');
+            $table->unsignedBigInteger('id_rhh');
             $table->tinyInteger('state');
             $table->timestamps();
         });
